requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
et-xmlfile==1.1.0
greenlet==3.0.3
idna==3.10
numpy==2.1.2
openpyxl==3.1.5
pandas==2.2.3
playwright==1.47.0
pyee==12.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
six==1.16.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
